(opm) move add from root to alpha
