+++ cmd/opm/alpha/cmd.go
@@ -1,6 +1,7 @@
 package alpha
 
 import (
+	"github.com/operator-framework/operator-registry/cmd/opm/alpha/add"
 	"github.com/operator-framework/operator-registry/cmd/opm/alpha/bundle"
 	"github.com/spf13/cobra"
 )
@@ -12,6 +13,6 @@ func NewCmd() *cobra.Command {
 		Short:  "Run an alpha subcommand",
 	}
 
-	runCmd.AddCommand(bundle.NewCmd())
+	runCmd.AddCommand(bundle.NewCmd(), add.NewCmd())
 	return runCmd
 }
